server/requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
chromedriver-autoinstaller==0.4.0
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.6.2
exceptiongroup==1.1.0
fastapi==0.94.0
h11==0.14.0
idna==3.4
outcome==1.2.0
pydantic==1.10.6
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.26.0.post1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
wsproto==1.2.0
